REFACTOR: adding user story to limit bikes to a default max of 20
Causes a number of issues when adding a default capacity to the docking station,
the number now requires there to be multiple objects saved, therefore decided
on an array. Have also had to initialize that array, and change the code to accomodate
that new object, which causes multiple tests to be redundant.