ADD/REFACTOR: double the bike object in the docking station and broken method on bike.
To ensure encapsulation i doubled the bike object in the docking station tests, was
not happy that we had tightly joined the two classes so tried to encapsulate. Have also
decided to start work on the bike class, with a broken method, which required me to
change the tests on the bike class - from working method, to the new broken method.