ADD/REFACTOR: broken method on Bike, and ability to raise error
various changes here mainly aligned to the case of the bike not being released
if it is in a broken state. Have considred just the last bike in the
array at this time, could code it better to loop over all bikes to find one
that isnt broken, but thats not what is required to pass the minimal test.